+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+  * sprop-max-don-diff SDP parameter #885
   * Reduce number of RTP header extensions sent on every packet #880
   * Add RTP-level keyframe detection for VP8, VP9, AV1, H264 and H265 #869
   * VP9 switch packetizer to non-flexible mode (F=0) #868